# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/course": {
      "post": {
        "operationId": "CourseController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Course"
        ]
      },
      "get": {
        "operationId": "CourseController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Course"
        ]
      }
    },
    "/api/course/{id}": {
      "get": {
        "operationId": "CourseController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Course"
        ]
      },
      "patch": {
        "operationId": "CourseController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Course"
        ]
      },
      "delete": {
        "operationId": "CourseController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Course"
        ]
      }
    },
...
    "/api/group/{id}": {
      "get": {
        "operationId": "GroupController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Group"
        ]
      },
      "patch": {
        "operationId": "GroupController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Group"
        ]
      },
      "delete": {
        "operationId": "GroupController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Group"
        ]
      }
    },
    "/api/branch": {
      "post": {
        "operationId": "BranchController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBranchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Branch"
        ]
      },
      "get": {
        "operationId": "BranchController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Branch"
        ]
      }
    },
    "/api/branch/{id}": {
      "get": {
        "operationId": "BranchController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Branch"
        ]
      },
      "patch": {
        "operationId": "BranchController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBranchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Branch"
        ]
      },
      "delete": {
        "operationId": "BranchController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Branch"
        ]
      }
    },
...
    "/api/district": {
      "post": {
        "operationId": "DistrictController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDistrictDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "District"
        ]
      },
      "get": {
        "operationId": "DistrictController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "District"
        ]
      }
    },
    "/api/district/{id}": {
      "get": {
        "operationId": "DistrictController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "District"
        ]
      },
      "patch": {
        "operationId": "DistrictController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDistrictDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "District"
        ]
      },
      "delete": {
        "operationId": "DistrictController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "District"
        ]
      }
    },
...
    "/api/group-student": {
      "post": {
        "operationId": "GroupStudentController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupStudentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "GroupStudent"
        ]
      },
      "get": {
        "operationId": "GroupStudentController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "GroupStudent"
        ]
      }
    },
    "/api/group-student/{id}": {
      "get": {
        "operationId": "GroupStudentController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "GroupStudent"
        ]
      },
      "patch": {
        "operationId": "GroupStudentController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupStudentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "GroupStudent"
        ]
      },
      "delete": {
        "operationId": "GroupStudentController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "GroupStudent"
        ]
      }
    },
...
    "/api/attendance": {
      "post": {
        "operationId": "AttendanceController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAttendanceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Attendance"
        ]
      },
      "get": {
        "operationId": "AttendanceController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Attendance"
        ]
      }
    },
    "/api/attendance/{id}": {
      "get": {
        "operationId": "AttendanceController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Attendance"
        ]
      },
      "patch": {
        "operationId": "AttendanceController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAttendanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Attendance"
        ]
      },
      "delete": {
        "operationId": "AttendanceController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Attendance"
        ]
      }
    },
...
    "/api/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/signin": {
      "post": {
        "operationId": "AuthController_signin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SigninDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Documentation",
    "description": "The API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateCourseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Nodejs, React, Nextjs, Typescript",
            "description": "Course name"
          },
          "description": {
            "type": "string",
            "example": "Nodejs, React, Nextjs, Typescript",
            "description": "Course description"
          },
          "capacity": {
            "type": "number",
            "example": 20,
            "description": "Maximum number of students"
          },
          "duration": {
            "type": "number",
            "example": 3,
            "description": "Course duration in months"
          },
          "lesson_quantity": {
            "type": "number",
            "example": 36,
            "description": "Total number of lessons"
          },
          "price": {
            "type": "number",
            "example": 1000000,
            "description": "Total course price"
          },
          "price_per_lesson": {
            "type": "number",
            "example": 30000,
            "description": "Price per lesson"
          }
        },
        "required": [
          "name",
          "description",
          "capacity",
          "duration",
          "lesson_quantity",
          "price",
          "price_per_lesson"
        ]
      },
      "UpdateCourseDto": {
        "type": "object",
        "properties": {}
      },
...
      "UpdateGroupDto": {
        "type": "object",
        "properties": {}
      },
      "CreateBranchDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Branch name",
            "example": "Branch A"
          },
          "region_id": {
            "type": "number",
            "description": "Region ID",
            "example": 1
          },
          "district_id": {
            "type": "number",
            "description": "District ID",
            "example": 1
          },
          "location": {
            "type": "string",
            "description": "Location",
            "example": "40.123456, 71.123456"
          }
        },
        "required": [
          "name",
          "region_id",
          "district_id",
          "location"
        ]
      },
      "UpdateBranchDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateDistrictDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "District name",
            "example": "Toshkent"
          },
          "region_id": {
            "type": "number",
            "description": "Region ID",
            "example": 1
          }
        },
        "required": [
          "name",
          "region_id"
        ]
      },
      "UpdateDistrictDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateGroupStudentDto": {
        "type": "object",
        "properties": {
          "group_id": {
            "type": "number",
            "description": "Group ID",
            "example": 1
          },
          "student_id": {
            "type": "number",
            "description": "Student ID",
            "example": 1
          }
        },
        "required": [
          "group_id",
          "student_id"
        ]
      },
      "UpdateGroupStudentDto": {
        "type": "object",
        "properties": {
          "group_id": {
            "type": "number",
            "description": "Group ID",
            "example": 1
          },
          "student_id": {
            "type": "number",
            "description": "Student ID",
            "example": 1
          }
        }
      },
...
      "CreateAttendanceDto": {
        "type": "object",
        "properties": {
          "student_id": {
            "type": "number",
            "description": "Student ID",
            "example": 1
          },
          "lesson_id": {
            "type": "number",
            "description": "Lesson ID",
            "example": 1
          },
          "date": {
            "type": "string",
            "description": "Attendance date",
            "example": "2025-11-11"
          }
        },
        "required": [
          "student_id",
          "lesson_id"
        ]
      },
      "UpdateAttendanceDto": {
        "type": "object",
        "properties": {
          "student_id": {
            "type": "number",
            "description": "Student ID",
            "example": 1
          },
          "lesson_id": {
            "type": "number",
            "description": "Lesson ID",
            "example": 1
          },
          "date": {
            "type": "string",
            "description": "Attendance date",
            "example": "2025-11-11"
          },
          "is_present": {
            "type": "boolean",
            "description": "Is present",
            "example": true
          }
        },
        "required": [
          "is_present"
        ]
      },
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John",
            "description": "User first name"
          },
          "surname": {
            "type": "string",
            "example": "Doe",
            "description": "User last name"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com",
            "description": "User email"
          },
          "phone": {
            "type": "number",
            "example": 998901234567,
            "description": "User phone number"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "User password"
          },
          "gender": {
            "type": "string",
            "example": "Male",
            "description": "User gender"
          },
          "xp": {
            "type": "number",
            "example": 100,
            "description": "Experience points"
          },
          "is_learning": {
            "type": "boolean",
            "example": true,
            "description": "Is learning"
          },
          "district_id": {
            "type": "number",
            "example": 1,
            "description": "District ID"
          },
          "job": {
            "type": "string",
            "example": "Teacher",
            "description": "Job title"
          },
          "hire_date": {
            "type": "string",
            "example": "2024-01-01",
            "description": "Hire date"
          },
          "salary": {
            "type": "number",
            "example": 1000,
            "description": "Salary"
          },
          "is_fired": {
            "type": "boolean",
            "example": false,
            "description": "Is fired"
          }
        },
        "required": [
          "name",
          "surname",
          "phone",
          "password"
        ]
      },
      "SigninDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "User ID",
            "example": 1
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "password123"
          }
        },
        "required": [
          "id",
          "password"
        ]
      },
...
    }
  }
}